visualize the subclass weight regression with a continuous covariate

visualize the subclass weight regression with a continuous covariate

Usage

plot_subwt_regression(
  DIR_NPLCM,
  stratum_bool,
  case = 0,
  slice = 1,
  truth = NULL,
  RES_NPLCM = NULL
)

Arguments

DIR_NPLCM

File path to the folder containing posterior samples

stratum_bool

a vector of TRUE/FALSE with TRUE indicating the rows of subjects to include

case

1 for plotting cases, 0 for plotting controls; default to 0.

slice

integer; specifies which slice of bronze-standard data to visualize; Default to 1.

truth

a list of truths computed from true parameters in simulations; elements: Eti, FPR, PR_case,TPR; All default to NULL in real data analyses. Currently only works for one slice of bronze-standard measurements (in a non-nested model).

truth_subwt matrix of # of rows = # of subjects, # columns: number of true subclasses

RES_NPLCM

pre-read res_nplcm; default to NULL.

Value

A figure of subclass regression curves

See also